backgrounds for web; names refactor
