+++ src/actions.js
@@ -5,7 +5,7 @@ const db = firebase.database();
 export const SET_ACTIVE_MENU_ITEM = 'SET_ACTIVE_MENU_ITEM';
 export const GET_PRODUCTS = 'GET_PRODUCTS';
 
-export const setActiveMenuList = item => ({type: SET_ACTIVE_MENU_ITEM, payload: item});
+export const setActiveMenuItem = item => ({type: SET_ACTIVE_MENU_ITEM, payload: item});
 export const getProducts = products => ({type: GET_PRODUCTS, payload: products});
 
 export function getProductsThunk() {
