+++ src/constants/routes.js
@@ -5,7 +5,7 @@ export const LANDING = '/';
 export const SIGN_UP = '/rejestracja';
 export const SIGN_IN = '/logowanie';
 export const INFO = '/informacje';
-export const DASHBOARD = '/panel';
+export const DASHBOARD = '/pulpit';
 export const SETTINGS = '/ustawienia';
 export const STEP_1 = '/ustawienia/krok-1';
 export const STEP_2 = '/ustawienia/krok-2';
