+++ src/styles/global.css
@@ -280,5 +280,3 @@ body {
     opacity: 0.75; }
   .input:hover {
     border-color: #1e656d; }
-
-/*# sourceMappingURL=global.css.map */
